+++ AmericanExpress/src/main/java/com/braintreepayments/api/americanexpress/AmericanExpressGetRewardsBalanceCallback.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.americanexpress
 
 /**
  * Callback for receiving result of