+++ AmericanExpress/src/main/java/com/braintreepayments/api/americanexpress/AmericanExpressResult.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.americanexpress
 
 /**
  * Result of fetching American Express rewards balance